src/main/java/com/woowacourse/matzip/application/RestaurantService.java
package com.woowacourse.matzip.application;

import com.woowacourse.matzip.application.response.RestaurantResponse;
import com.woowacourse.matzip.domain.restaurant.RestaurantRepository;
import java.util.List;
import java.util.stream.Collectors;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@Transactional(readOnly = true)
public class RestaurantService {

    private final RestaurantRepository restaurantRepository;

    public RestaurantService(final RestaurantRepository restaurantRepository) {
        this.restaurantRepository = restaurantRepository;
    }

    public List<RestaurantResponse> findByCampusId(final Long campusId) {
        return restaurantRepository.findByCampusId(campusId)
                .stream()
                .map(RestaurantResponse::from)
                .collect(Collectors.toList());
    }
}
